requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.25.0
